+++ NEWS.md
@@ -1,3 +1,9 @@
+# tidylavaan 1.2.0-0  
+
+  - Reduce dependacies.
+  - Added option to pass args to lavaan methods.
+  - Added option to include variances.
+
 # tidylavaan 1.1.0-0  
   
   - Support for unstandardized parameters